use group page as root for single_group_mode
